pyproject.toml
[tool.poetry]
name = "enterprise-log-analyzer"
version = "0.1.0"
description = "FastAPI-based log ingestion & analysis"
authors = ["Roberto roberto.hordoan@computacenter.com"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.110.0"
uvicorn = { extras = ["standard"], version = ">=0.35.0" }
sqlalchemy = "^2.0"
alembic = "^1.12.0"
pydantic = "^2.5"
redis = "^4.4"
aiofiles = "^23.1"
asyncpg = "^0.30.0"
ruff = "^0.13.0"
pydantic-settings = "^2.10.1"
chromadb = ">=0.5.5"
openai = ">=1.40.0"
sentence-transformers = ">=2.2.2"
# numpy and scikit-learn are heavy; only needed when using sentence-transformers locally
# keep them optional to reduce container size and dependency pain
ollama = "^0.5.4"
pyyaml = "^6.0.2"
httpx = ">=0.27.0"
puresnmp = ">=2.0.0"
opentelemetry-sdk = ">=1.26.0"
opentelemetry-exporter-otlp = ">=1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
httpx = ">=0.27.0"
fakeredis = "^2.0.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
